Organize test directory
hogehoge
